refactor Confirmation Exception
