+++ src/main/java/com/khutircraftubackend/confirm/exception/ConfirmationException.java
@@ -1,12 +1,12 @@
-package com.khutircraftubackend.confirm;
+package com.khutircraftubackend.confirm.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class CustomConfirmException extends RuntimeException{
+public class ConfirmationException extends RuntimeException {
 
-    public CustomConfirmException(String message){
+    public ConfirmationException(String message) {
         super (message);
     }
 }
